requirements.txt
PyPDF2
sentence-transformers
langchain
pinecone-client
python-dotenv
keras-nlp
keras>=3
tensorflow